Remove NDex functionality.
Since I created a project called NDex which contains the code for
working with sub-lists, I removed the duplicate code from this project.

I also removed the Converter and the PriorityQueue. PriorityQueue is
implemented in terms of NDex and not heavily used. The Converter class
doesn't belong in a collection library.